Added working weaponry, but the engine is a little glitchy...

+++ code/weaponClass.moon
@@ -1,4 +1,5 @@
 import random from love.math
+import ceil from math
 
 export newWeaponBox = -> --player required as an OBJECT to fire
 	return weaponsList[random(1,#weaponsList)] --gives player a class which can be used to create new projectiles
@@ -10,11 +11,20 @@ export newWeaponBox = -> --player required as an OBJECT to fire
 -- Characters can find out what weapon they have by checking Weapon.__name (which equals to its class)
 
 export class weapon  --main class, here pretty much just for inheritance
+	new: =>
+		@x=0
+		@y=0
+		@xVel=0
+		@yVel=0
 	inherit: (player) =>
 		@owner=player
 		@gravity=2
 		@floor=player.floor
-		switch @facing
+		@x=0
+		@y=0
+		@xVel=0
+		@yVel=0
+		switch player.facing
 			when "left"
 				@x=player.x-1
 				@y=player.y
@@ -37,14 +47,6 @@ export class weapon  --main class, here pretty much just for inheritance
 				@yVel=1
 		@above=20 --pix above ground
 		@zVel=0.5 --Vels == velocities, added on to existing forces. to add to confusion: x and y vels are in BLOCKS, zVel is in PIXELS. also remember this is only run every so often (alright, probably about 10 times a second or something)
-
-export class shotgun extends weapon
-	ammo: 5 --decrease on shot, please
-	new: (player) =>
-		@inherit player
-		--xVel,yVels are fine, should only be changed for drills and other stuff™
-		@drag=0.2
-		@damage=((@xVel+@yVel+@zVel)/3)*8 --should change, check per frame
 	run: =>
 		@x+=@xVel
 		@y+=@yVel
@@ -52,11 +54,32 @@ export class shotgun extends weapon
 		@yVel-=(@yVel*@drag)
 		@zVel+=@gravity
 		@above-=@zVel
-		@=nil if not @collision!
+		@damage=((@xVel+@yVel+@zVel)/3)*8 --should change, check per frame
+		return false if not @collision!
 		if @above<0
 			@floor+=1
-			@above=280-@above --CHECK 280 IT'S DOING WEIRD THINGS TO THE FALLING
-	collision: =>
+			@above=280-@above
+		return true
+	collision: =>  --seems to not be running? or it returns false verrrryyy quickly
+		return false if @x>5 or @x<1 or @y<1 or @x>5 or @floor<1 or @floor>mapSize
+		if @above<20 and map[ceil @floor][ceil @x][ceil @y]~=0
+			print @damage
+			if @damage>2  --needs to be replaced with a call to a Block class, extended by type, or even just a reference table
+				map[ceil @floor][ceil @x][ceil @y]=0
+				@xVel/=2
+				@yVel/=2
+				@zVel/=2
+			else return false
+		return true --should be at end
 		--check to see if a block is hit, (if @above<0 deal damage), subtract damage and dramatically subtract from x y z velocities, if damage is too low or x+y+z==0 then return false and kill this weapon (set to nil)
+
+export class shotgun extends weapon
+	ammo: 5 --decrease on shot, please
+	begin: (player) =>
+		@inherit player
+		--xVel,yVels are fine, should only be changed for drills and other stuff™
+		@drag=0.2
+		@above=60
+		@damage=((@xVel+@yVel+@zVel)/3)*8 --should change, check per frame
 
 export weaponsList = {shotgun} --last just to make sure
